requirements.txt
streamlit
selenium
webdriver-manager
python-dotenv
requests
mysql-connector-python
sqlalchemy
pymysql
pandas